+++ api/tests/signup/signup-phone.test.ts
@@ -1,16 +1,14 @@
 import { test, expect } from '@playwright/test'
-import { OTP } from '../../../shared/utils/seed-config'
 import { phoneSignupProfile } from '../../helpers/profiles/phone'
 import {
   firebaseSignIn,
   firebaseRefreshToken as firebaseRefreshTokenAPI,
 } from '../../helpers/firebase'
 import { setupSeedTeardown } from '../../helpers/test-setup'
-import { parseResponse } from '../../../shared/utils/response'
-import { createPin, getProfile, logout } from '../../helpers/signup-flow'
-import { OtpRequestSchema, OtpVerifySchema } from '../../schema/signup.schema'
-import { endpoints } from '../../../shared/endpoints'
-import { DEFAULT_REQUEST_HEADERS, PHONE_VERIFY_PARAMS } from '../../helpers/request'
+import { createPin } from '../../helpers/pin-api'
+import { getProfile } from '../../helpers/profile-api'
+import { logout } from '../../helpers/auth-api'
+import { requestPhoneOtp, verifyPhoneOtp } from '../../helpers/phone-signup-api'
 
 test.describe('Signup by Phone', () => {
   const { beforeEach, afterEach, getContext } = setupSeedTeardown(phoneSignupProfile)
@@ -30,28 +28,11 @@ test.describe('Signup by Phone', () => {
       let idTokenPostPin: string
 
       await test.step('Request OTP', async () => {
-        const payload = { phone }
-        const response = await request.post(endpoints.signup.requestOtp, {
-          data: payload,
-          headers: DEFAULT_REQUEST_HEADERS,
-        })
-        const parsed = await parseResponse(response, OtpRequestSchema, 'Request OTP', 200, payload)
-        expect(parsed.is_signup).toBe(false)
-        expect(parsed.next_state).toBe('signup.phone.verify')
-        refCode = parsed.verification_info.ref_code
+        refCode = await requestPhoneOtp(request, phone)
       })
 
       await test.step('Verify OTP', async () => {
-        const payload = { phone, ref_code: refCode, code: OTP }
-        const response = await request.post(endpoints.signup.verifyOtp, {
-          params: PHONE_VERIFY_PARAMS,
-          data: payload,
-          headers: DEFAULT_REQUEST_HEADERS,
-        })
-        const parsed = await parseResponse(response, OtpVerifySchema, 'Verify OTP', 200, payload)
-        expect(parsed.is_signup).toBe(true)
-        expect(parsed.next_state).toBe('user.profile')
-        firebaseCustomToken = parsed.verification_info.token
+        firebaseCustomToken = await verifyPhoneOtp(request, phone, refCode)
       })
 
       await test.step('Firebase sign in with custom token', async () => {
@@ -78,6 +59,10 @@ test.describe('Signup by Phone', () => {
         expect(body.profile.phone).toBe(phone)
         expect(body.profile.has_pincode).toBe(true)
         expect(body.profile.signup_at).not.toBeNull()
+      })
+
+      await test.step('Logout', async () => {
+        await logout(request, idTokenPostPin)
       })
     }
   )
